Rename layer to AvalancheBulletinsLayer
