+++ src/components/yeti/map-layers/AvalancheBulletinsLayer.vue
